requirements.txt
pyOpenSSL==23.2.0
lxml==4.9.2
requests==2.31.0
zeep==4.2.1
